+++ test/eslint/rules/event-const-desc.js
@@ -24,7 +24,19 @@ module.exports = extendClassDeclarations({
 			` * Fired with: nothing`,
 			` * @const {string}`,
 			` */`,
-			`Klass.EVENT_CHANGE = 'change';`
+			`export const EVENT_ERROR = 'error';`,
+			``,
+			`/**`,
+			` * Fired with: nothing`,
+			` * @const {string}`,
+			` */`,
+			`Klass.EVENT_CHANGE = 'change';`,
+			``,
+			`/**`,
+			` * Fired with: nothing`,
+			` * @const {string}`,
+			` */`,
+			`EVENT_DONE = 'done';`
 		)
 	}],
 	invalid: [{
@@ -41,11 +53,23 @@ module.exports = extendClassDeclarations({
 			`/**`,
 			` * @const {string}`,
 			` */`,
-			`Klass.EVENT_CHANGE = 'change';`
+			`export const EVENT_ERROR = 'error';`,
+			``,
+			`/**`,
+			` * @const {string}`,
+			` */`,
+			`Klass.EVENT_CHANGE = 'change';`,
+			``,
+			`/**`,
+			` * @const {string}`,
+			` */`,
+			`EVENT_DONE = 'done';`
 		),
 		errors: errors(
 			`Event constant "EVENT_CLICK" has no "Fired with: ..." description.`,
-			`Event constant "EVENT_CHANGE" has no "Fired with: ..." description.`
+			`Event constant "EVENT_ERROR" has no "Fired with: ..." description.`,
+			`Event constant "EVENT_CHANGE" has no "Fired with: ..." description.`,
+			`Event constant "EVENT_DONE" has no "Fired with: ..." description.`
 		)
 	}]
 });