+++ test/eslint/rules/event-const-value.js
@@ -15,7 +15,17 @@ module.exports = extendClassDeclarations({
 			`/**`,
 			` * @const {string}`,
 			` */`,
-			`Klass.EVENT_STATE_CHANGE = 'state-change';`
+			`export const EVENT_ERROR = 'error';`,
+			``,
+			`/**`,
+			` * @const {string}`,
+			` */`,
+			`Klass.EVENT_STATE_CHANGE = 'state-change';`,
+			``,
+			`/**`,
+			` * @const {string}`,
+			` */`,
+			`EVENT_HOVER = 'hover';`
 		)
 	}],
 	invalid: [{
@@ -32,11 +42,23 @@ module.exports = extendClassDeclarations({
 			`/**`,
 			` * @const {string}`,
 			` */`,
-			`Klass.EVENT_STATE_CHANGE = 3;`
+			`export const EVENT_ERROR = 3;`,
+			``,
+			`/**`,
+			` * @const {string}`,
+			` */`,
+			`Klass.EVENT_STATE_CHANGE = null;`,
+			``,
+			`/**`,
+			` * @const {string}`,
+			` */`,
+			`EVENT_HOVER = undefined;`
 		),
 		errors: errors(
 			`Event constant "EVENT_RIGHT_CLICK" value should be a string.`,
-			`Event constant "EVENT_STATE_CHANGE" value should be a string.`
+			`Event constant "EVENT_ERROR" value should be a string.`,
+			`Event constant "EVENT_STATE_CHANGE" value should be a string.`,
+			`Event constant "EVENT_HOVER" value should be a string.`
 		)
 	}, {
 		code: concat(
@@ -46,41 +68,40 @@ module.exports = extendClassDeclarations({
 			`        * @const {string}`,
 			`        */`,
 			`       this.EVENT_RIGHT_CLICK = 'Right-Click';`,
-			`   }`,
-			`}`,
 			``,
-			`/**`,
-			` * @const {string}`,
-			` */`,
-			`Klass.EVENT_STATE_CHANGE = 'hover:state:change';`
-		),
-		errors: errors(
-			`Event constant "EVENT_RIGHT_CLICK" should consist of latinic lowercase letters (a-z) ` +
-			`or dash sign (-), but "R" given.`,
-			`Event constant "EVENT_STATE_CHANGE" should consist of latinic lowercase letters (a-z) ` +
-			`or dash sign (-), but ":" given.`
-		)
-	}, {
-		code: concat(
-			`class Klass {`,
-			`   constructor() {`,
 			`       /**`,
 			`        * @const {string}`,
 			`        */`,
-			`       this.EVENT_RIGHT_CLICK = 'right_click';`,
+			`       this.EVENT_LEFT_CLICK = 'left_click';`,
 			`   }`,
 			`}`,
 			``,
 			`/**`,
 			` * @const {string}`,
 			` */`,
-			`Klass.EVENT_STATE_CHANGE = '3';`
+			`export const EVENT_ERROR = 'error!';`,
+			``,
+			`/**`,
+			` * @const {string}`,
+			` */`,
+			`Klass.EVENT_STATE_CHANGE = 'state:change';`,
+			``,
+			`/**`,
+			` * @const {string}`,
+			` */`,
+			`EVENT_HOVER = 'hover?';`
 		),
 		errors: errors(
-			`Event constant "EVENT_RIGHT_CLICK" should consist of latinic lowercase letters (a-z) ` +
+			`Event constant "EVENT_RIGHT_CLICK" should consist of latin lowercase letters (a-z) ` +
+			`or dash sign (-), but "R" given.`,
+			`Event constant "EVENT_LEFT_CLICK" should consist of latin lowercase letters (a-z) ` +
 			`or dash sign (-), but "_" given.`,
-			`Event constant "EVENT_STATE_CHANGE" should consist of latinic lowercase letters (a-z) ` +
-			`or dash sign (-), but "3" given.`
+			`Event constant "EVENT_ERROR" should consist of latin lowercase letters (a-z) ` +
+			`or dash sign (-), but "!" given.`,
+			`Event constant "EVENT_STATE_CHANGE" should consist of latin lowercase letters (a-z) ` +
+			`or dash sign (-), but ":" given.`,
+			`Event constant "EVENT_HOVER" should consist of latin lowercase letters (a-z) ` +
+			`or dash sign (-), but "?" given.`
 		)
 	}]
 });